🛣📒 Checkpoint ./ayarlar.json:23186552/96 ./bot.js:23186552/68 ./komutlar/botgörüldü.js:23186552/2154 ./komutlar/botekle-reddet.js:23186552/3138 ./komutlar/botekle-onayla.js:23186552/2078 ./komutlar/botekle-ekle.js:23186552/4003 ./komutlar/botkapalı.js:23186552/42 ./util/eventLoader.js:23186552/379 ./komutlar/botat.js:23186552/10

+++ komutlar/botekle-onayla.js
@@ -1,28 +1,39 @@
-const Discord = require('discord.js');
-
+const Discord = require("discord.js");
 
 exports.run = function(client, message, args) {
-  
-  if (!message.member.hasPermission("ADMINISTRATOR")) return message.channel.send(`:no_entry: Bu komutu kullanabilmek için "\`Yönetici\`" yetkisine sahip olmalısın.`);
-	let lrowsisim = args[0]
-  let sahip = args[1]
-	let log = "742455429677842533" // bot eklendi / onaylandı / reddedildi kanalı
-	
-	if (!lrowsisim) return message.channel.send(`:no_entry: Botun idsini yazmalısın.`).then(msg => msg.delete(10000))
-  message.delete()
-		client.channels.get(log).send(`:white_check_mark: <@${sahip}> adlı kişinin <@${lrowsisim}> adlı botu onaylandı.`);
-		message.channel.send(`:white_check_mark: Botu onayladınız.`).then(msg => msg.delete(10000))
+  if (!message.member.hasPermission("ADMINISTRATOR"))
+    return message.channel.send(
+      `:no_entry: Bu komutu kullanabilmek için "\`Yönetici\`" yetkisine sahip olmalısın.`
+    );
+  let narcoscodisim = args[0];
+  let sahip = args[1];
+  let log = "742455429677842533"; // bot eklendi / onaylandı / reddedildi kanalı
+
+  if (!narcoscodisim)
+    return message.channel
+      .send(`:no_entry: Botun idsini yazmalısın.`)
+      .then(msg => msg.delete(10000));
+  message.delete();
+  client.channels
+    .get(log)
+    .send(
+      `:white_check_mark: <@${sahip}> adlı kişinin <@${narcoscodisim}> adlı botu onaylandı.`
+    );
+  message.channel
+    .send(`:white_check_mark: Botu onayladınız.`)
+    .then(msg => msg.delete(10000));
 };
 
 exports.conf = {
   enabled: true,
   guildOnly: false,
-  aliases: ['bot-onayla', 'onayla'],
+  aliases: ["bot-onayla", "onayla"],
   permLevel: 0
 };
 
 exports.help = {
-  name: 'botonayla', 
+  name: "botonayla",
   description: "Sunucuya eklenen botu onaylar.",
-  usage: 'botonayla <bot ismi>'
-};6
+  usage: "botonayla <bot ismi>"
+};
+6;
